// src/emission/services/waste-disposal/waste-disposal.enum.ts
export enum WasteDisposalMethod {

    RE_USE = "Re-use",
    OPEN_LOOP = "Open loop",
    CLOSED_LOOP = "Closed-loop",
    COMBUSION = "Combusion",
    COMPOSTING = "Composting" ,
    LANDFILL = "Landfill",
    ANAEROBIC_DIGESTION = "Anaerobic digestion",
    PIGGERY_FEEDING = "Piggery Feeding",
    INCINERATION = "Incineration",
}




export enum WasteDisposal {

    AVERAGE_CONSTRUCTION = "AVERAGE_CONSTRUCTION" ,
    TYRES = "TYRES",
    WOODS = "WOODS",
    GLASS = "GLASS",
    MUNICIPAL_WASTE = "MUNICIPAL_WASTE",
    WEEE = "WEEE",
    WEEE_FRIDGES_AND_FREEZERS = "WEEE_FRIDGES_AND_FREEZERS",
    BATTERIES = "BATTERIES",
    PLASTIC = "PLASTIC",
    SOILS = "SOILS",
    WOOD = "WOOD",
    BOOKS = "BOOKS",
    CLOTHING = "CLOTHING",
    COMMERCIAL_AND_INDUSTRIAL_WASTE = "COMMERCIAL_AND_INDUSTRIAL_WASTE",
    METAL = "METAL",
    PAPER_AND_BOARD = "PAPER_AND_BOARD",
    ORGANIC_FOOD_AND_DRINK_WASTE = "ORGANIC_FOOD_AND_DRINK_WASTE",
    ORGANIC_GARDEN_WASTE = "ORGANIC_GARDEN_WASTE",
    ORGANIC_MIXED_FOOD_AND_GARDEN_WASTE = "ORGANIC_MIXED_FOOD_AND_GARDEN_WASTE",
    WASTE= "WASTE",
    PAPER_CARDBOARD= "PAPER_CARDBOARD",
    TEXTILE= "TEXTILE",
    FOOD_WASTE="FOOD_WASTE",
    GARDEN_AND_PARK_WASTE= "GARDEN_AND_PARK_WASTE",
    NAPPIES="NAPPIES",
    RUBBER_AND_LEATHER=  "RUBBER_AND_LEATHER",
    PLASTICS="PLASTICS",
    OTHER_WASTES="OTHER_WASTES",
    DOMESTIC="DOMESTIC",
    INDUSTRIAL = "INDUSTRIAL",



    AGGREGATES = "AGGREGATES",
    ASBESTOS = "ASBESTOS", 
    ASPHALT = "ASPHALT",
    BRICKS = "BRICKS",
    CONCRETE = "CONCRETE",
    INSULATION = "INSULATION",
 
    MINERAL_OIL = "MINERAL_OIL",
    PLASTERBOARD = "PLASTERBOARD",
    HOUSEHOLD_RESIDUAL_WASTE= "HOUSEHOLD_RESIDUAL_WASTE",


    WEEE_LARGE= "WEEE_LARGE",
    WEEE_MIXED = "WEEE_MIXED",
    WEEE_SMALL = "WEEE_SMALL",

    METAL_ALUMINIUMCANS_FOIL = "METAL_ALUMINIUMCANS_FOIL",
    METAL_MIXED_CANS = "METAL_MIXED_CANS",
    METAL_SCRAP_METAL = "METAL_SCRAP_METAL",
    METAL_STEEL_CANS = "METAL_STEEL_CANS",


    PLASTICS_AVERAGE_PLASTICS = "PLASTICS_AVERAGE_PLASTICS",
    PLASTICS_AVERAGE_PLASTIC_FILM ="PLASTICS_AVERAGE_PLASTIC_FILM",
    PLASTICS_AVERAGE_PLASTIC_RIGID = "PLASTICS_AVERAGE_PLASTIC_RIGID",
    PLASTICS_HDPE = "PLASTICS_HDPE",
    PLASTICS_LDPE_LLDPE = "PLASTICS_LDPE_LLDPE",
    PLASTICS_PET = "PLASTICS_PET",
    PLASTICS_PP = "PLASTICS_PP",
    PLASTICS_PS = "PLASTICS_PS",
    PLASTICS_PVC = "PLASTICS_PVC",
    

    PAPER_BOARD_BOARD = "PAPER_BOARD_BOARD",
    PAPER_BOARD_MIXED =  "PAPER_BOARD_MIXED",
    PAPER_BOARD_PAPER = "PAPER_BOARD_PAPER",




}

export enum WasteActivities {
    BIOLOGICAL_TREATMENT = "BIOLOGICAL_TREATMENT",
    WASTE_INCINERATION = "WASTE_INCINERATION",
    OPEN_BURNING = "OPEN_BURNING",
    DOMESTIC_WASTEWATER = "DOMESTIC_WASTEWATER",
    INDUSTRIAL_WASTEWATER = "INDUSTRIAL_WASTEWATER",
    OTHER = "OTHER_ACTIVITIES_BASED_ON_DEFRA",
    INCINERATION = "INCINERATION"
}